+++ lib/html_sigs.mli
@@ -180,6 +180,7 @@ module type T = sig
   val a_oncanplay : Xml.event_handler -> [> | `OnCanPlay] attrib
   val a_oncanplaythrough : Xml.event_handler -> [> | `OnCanPlayThrough] attrib
   val a_onchange : Xml.event_handler -> [> | `OnChange] attrib
+  val a_onclose : Xml.event_handler -> [> | `OnClose] attrib
   val a_ondurationchange : Xml.event_handler -> [> | `OnDurationChange] attrib
   val a_onemptied : Xml.event_handler -> [> | `OnEmptied] attrib
   val a_onended : Xml.event_handler -> [> | `OnEnded] attrib
@@ -768,6 +769,10 @@ module type T = sig
   val blockquote :
     ([< | blockquote_attrib], [< | blockquote_content_fun], [> | blockquote])
       star
+
+  val dialog :
+    ([< | dialog_attrib], [< | dialog_content_fun], [> | dialog]) star
+  (** @see <https://developer.mozilla.org/en-US/docs/Web/HTML/Element/dialog> *)
 
   val div : ([< | div_attrib], [< | div_content_fun], [> | div]) star